requirements.txt
seaborn
pyyaml
wandb
networkx
scipy
sklearn
onnx
onnxsim
pydot
plotly
dotmap
onnxoptimizer<=0.2.7